+++ python_ms_core/__init__.py
@@ -1,0 +1,8 @@
+from .main import Core
+from core.queue import topic
+from core.storage.providers.azure import azure_storage_client, azure_file_entity
+from core.logger.logger import Logger
+
+
+class Core(object):
+    pass
