+++ python_ms_core/core/queue/providers/azure_queue_config.py
@@ -1,6 +1,6 @@
 import os
 from dotenv import load_dotenv
-from models.config import CoreConfig
+from ..models.config import CoreConfig
 
 load_dotenv()
