+++ python_ms_core/core/storage/models/config.py
@@ -1,0 +1,13 @@
+import os
+from dotenv import load_dotenv
+
+load_dotenv()
+
+
+class CoreConfig:
+    def __init__(self):
+        self.provider = os.environ.get('PROVIDER', 'Azure')
+
+    @staticmethod
+    def default():
+        return CoreConfig()
