+++ requirements.txt
@@ -6,6 +6,8 @@ certifi==2022.12.7
 cffi==1.15.1
 charset-normalizer==2.1.1
 cryptography==38.0.4
+gitdb==4.0.10
+GitPython==3.1.29
 idna==3.4
 isodate==0.6.1
 msrest==0.7.1
@@ -15,7 +17,7 @@ python-dotenv==0.21.0
 requests==2.28.1
 requests-oauthlib==1.3.1
 six==1.16.0
+smmap==5.0.0
 typing_extensions==4.4.0
 uamqp==1.6.3
 urllib3==1.26.13
-version==0.1.0
