--- src/__init__.py
@@ -1,15 +1,0 @@
-# import os
-# import sys
-#
-# PROJECT_ROOT = os.path.abspath(os.path.join(
-#     os.path.dirname(__file__),
-#     os.pardir)
-# )
-#
-# sys.path.insert(0, PROJECT_ROOT)
-# directories = ['utils']
-# index = 1
-# for dir in directories:
-#     dir_path = os.path.abspath(os.path.join(PROJECT_ROOT, dir))
-#     sys.path.insert(index, dir_path)
-#     index += 1
